# loadtest/dashboard/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Phone API 명세서",
    "description": "핸드폰 중복 방지 API 문서입니다.",
    "version": "v0.0.4"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/inventory/{userId}": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "인벤토리 조회",
        "description": "유저의 보유 아이템 목록 조회",
        "operationId": "getInventory",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserItemResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "유저 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "itemType": {
            "type": "string"
          },
          "slotType": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          },
          "assetUrl": {
            "type": "string"
          },
          "paymentType": {
            "type": "string"
          },
          "currencyPrice": {
            "type": "integer",
            "format": "int32"
          },
          "premiumPrice": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UserItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item": {
            "$ref": "#/components/schemas/ItemResponse"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}